+++ daycan-back/api/settings.gradle
