+++ daycan-back/api/src/main/java/com/daycan/logging/AopLogger.java
@@ -1,4 +1,4 @@
-package com.tmp.boiler.logging;
+package com.daycan.logging;
 
 import lombok.extern.slf4j.Slf4j;
 import org.aspectj.lang.ProceedingJoinPoint;
