+++ daycan-back/api/src/main/java/com/daycan/logging/MdcLoggingInterceptor.java
@@ -1,4 +1,4 @@
-package com.tmp.boiler.logging;
+package com.daycan.logging;
 
 
 import jakarta.servlet.http.HttpServletRequest;
